requirements.txt
annotated-types==0.5.0
anyio==4.0.0
click==8.1.7
fastapi==0.103.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pypng==0.20220715.0
python-dotenv==1.0.0
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
